assets/requirements/requirements-base.txt
# Apollova Base Requirements — pinned versions for stability
# Setup.exe will enforce these exact versions, uninstalling conflicting ones first.

# Audio processing
yt-dlp
pydub==0.25.1
librosa==0.10.2

# Transcription deps — openai-whisper and stable-ts are installed in a
# dedicated step AFTER torch with --no-deps, so their transitive
# dependencies must be listed here explicitly.
tiktoken
more-itertools
numba
tqdm

# Vocal separation (improves Whisper accuracy on music)
demucs==4.0.1

# Audio preprocessing (noise reduction + language detection)
noisereduce
soundfile
langdetect

# Faster transcription backend (optional — graceful fallback to standard whisper)
faster-whisper

# Lyrics
lyricsgenius==3.0.1
rapidfuzz==3.9.7
beautifulsoup4==4.12.3

# Image processing
colorthief==0.2.1
Pillow==10.4.0
requests==2.32.3

# Numeric (must be <2 for torch compatibility)
numpy==1.26.4

# Environment / config
python-dotenv==1.0.1

# Song discovery (YouTube search)
pytubefix

# GUI
PyQt6==6.7.1

# Mobile remote control server
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
qrcode[pil]==7.4.2
